==> projects/1.3 Practice_data_loading.ipynb <==
# **Hands-on Practice Lab: Importing Dataset - Laptops Pricing**

Estimated time needed: **20** minutes

In this lab, you will practice the process of loading and drawing basic insights on a dataset as learnt through the module. You are being provided with a fresh dataset on 'Laptop Pricing' which will be used for all the practice labs throughout the course.


# Objectives

After completing this lab you will be able to:

 - Import a dataset from a CSV file to a Pandas dataframe
 - Develop some basic insights about the dataset


# Setup


For this lab, we will be using the following libraries:

* `skillsnetwork` for downloading the daataset

*   [`pandas`](https://pandas.pydata.org/?utm_medium=Exinfluencer&utm_source=Exinfluencer&utm_content=000026UJ&utm_term=10006555&utm_id=NA-SkillsNetwork-Channel-SkillsNetworkCoursesIBMML0187ENSkillsNetwork31430127-2021-01-01) for managing the data.
*   [`numpy`](https://numpy.org/?utm_medium=Exinfluencer&utm_source=Exinfluencer&utm_content=000026UJ&utm_term=10006555&utm_id=NA-SkillsNetwork-Channel-SkillsNetworkCoursesIBMML0187ENSkillsNetwork31430127-2021-01-01) for mathematical operations.


### Importing Required Libraries


import pandas as pd
import numpy as np

This function will download the dataset we need.


import requests

def download_csv(file_url, save_path):
    """
    Downloads a CSV file from the provided URL and saves it to the specified local path.
    
    Parameters:
    - file_url (str): URL of the file to download.
    - save_path (str): Local path to save the downloaded file.
    """
    try:
        response = requests.get(file_url)
        response.raise_for_status()  # Check if the request was successful
        
        with open(save_path, 'wb') as file:
            file.write(response.content)
        
        print(f"File downloaded successfully and saved to {save_path}")
    
    except requests.exceptions.RequestException as e:
        print(f"Error downloading file: {e}")



file_url = "https://cf-courses-data.s3.us.cloud-object-storage.appdomain.cloud/IBMDeveloperSkillsNetwork-DA0101EN-Coursera/laptop_pricing_dataset_base.csv"
save_path = "laptops.csv"

download_csv(file_url, save_path)

Read data into dataframe

file_name="laptops.csv"
df = pd.read_csv(file_name)

 Note: User can simply **skip the steps above**, and simply use the URL directly in the `pandas.read_csv()` function. You can uncomment and run the statements in the cell below.


filepath = "https://cf-courses-data.s3.us.cloud-object-storage.appdomain.cloud/IBMDeveloperSkillsNetwork-DA0101EN-Coursera/laptop_pricing_dataset_base.csv"
df = pd.read_csv(filepath, header=None)

<h1> Task #1: </h1>
<h3>Load the dataset to a pandas dataframe named 'df'</h3>
Print the first 5 entries of the dataset to confirm loading.


# Write your code below and press Shift+Enter to execute.
df = pd.read_csv(file_name, header=None)
print(df.head())

<details><summary>Click here for solution</summary>

```python
df = pd.read_csv(file_name, header=None)
print(df.head())
```
</details>


<h1> Task #2: </h1>
<h3>Add headers to the dataframe</h3>
The headers for the dataset, in sequence, are "Manufacturer", "Category", "Screen", "GPU", "OS", "CPU_core",
"Screen_Size_inch", "CPU_frequency", "RAM_GB", "Storage_GB_SSD", "Weight_kg" and "Price".

Confirm insertion by printing the first 10 rows of the dataset.


# Write your code below and press Shift+Enter to execute.
df.columns = ["Manufacturer", "Category", "Screen", "GPU", "OS", "CPU_core", "Screen_Size_inch", "CPU_frequency", "RAM_GB", "Storage_GB_SSD", "Weight_kg" , "Price"]

df.head(10)

<details><summary>Click here for solution</summary>

```python
# create headers list
headers = ["Manufacturer", "Category", "Screen", "GPU", "OS", "CPU_core", "Screen_Size_inch", "CPU_frequency", "RAM_GB", "Storage_GB_SSD", "Weight_kg", "Price"]
df.columns = headers
print(df.head(10))
```
</details>


<h1> Task #3: </h1>
<h3>Replace '?' with 'NaN'</h3>
Replace the '?' entries in the dataset with NaN value, recevied from the Numpy package.


# Write your code below and press Shift+Enter to execute.
df.replace('?', np.nan, inplace = True)

<details><summary>Click here for solution</summary>

```python
df.replace('?',np.nan, inplace = True)
```
</details>


<h1> Task #4: </h1>
<h3>Print the data types of the dataframe columns</h3>
Make a note of the data types of the different columns of the dataset.


# Write your code below and press Shift+Enter to execute.
df.dtypes

<details><summary>Click here for solution</summary>

```python
print(df.dtypes)
```
</details>


<h1> Task #5: </h1>
<h3>Print the statistical description of the dataset, including that of 'object' data types.</h3>


# Write your code below and press Shift+Enter to execute.
df.describe(include = 'all')

df.describe()

<details><summary>Click here for solution</summary>

```python
print(df.describe(include='all'))
```
</details>


<h1> Task #6: </h1>
<h3>Print the summary information of the dataset.</h3>



print(df.info())

<!--## Change Log


<!--|Date (YYYY-MM-DD)|Version|Changed By|Change Description|
|-|-|-|-|
|2023-09-15|0.1|Abhishek Gagneja|Initial Version Created|
|2023-09-18|0.2|Vicky Kuo|Reviewed and Revised|--!>
